requirements.txt
PySide6
Pillow
cryptography
pyinstaller
